move native frame collection out of processor
